+++ dotnet/Apps/Database/Domain.Tests/Invoice/Salesinvoicetests.cs
@@ -1241,6 +1241,34 @@ namespace Allors.Database.Domain.Tests
         }
 
         [Fact]
+        public void ChangedBillToCustomerDeriveDerivedLocale()
+        {
+            var localeGb = new LocaleBuilder(this.Transaction)
+                .WithCountry(new Countries(this.Transaction).FindBy(this.M.Country.IsoCode, "GB"))
+                .WithLanguage(new Languages(this.Transaction).FindBy(this.M.Language.IsoCode, "en"))
+                .Build();
+            var localeSe = new LocaleBuilder(this.Transaction)
+                .WithCountry(new Countries(this.Transaction).FindBy(this.M.Country.IsoCode, "SE"))
+                .WithLanguage(new Languages(this.Transaction).FindBy(this.M.Language.IsoCode, "sv"))
+                .Build();
+
+            var customer1 = new OrganisationBuilder(this.Transaction).WithName("customer1").WithLocale(localeGb).Build();
+            var customer2 = new OrganisationBuilder(this.Transaction).WithName("customer2").WithLocale(localeSe).Build();
+            new CustomerRelationshipBuilder(this.Transaction).WithFromDate(this.Transaction.Now()).WithCustomer(customer1).Build();
+            new CustomerRelationshipBuilder(this.Transaction).WithFromDate(this.Transaction.Now()).WithCustomer(customer2).Build();
+
+            var invoice = new SalesInvoiceBuilder(this.Transaction).WithBillToCustomer(customer1).Build();
+            this.Derive();
+
+            Assert.Equal(localeGb, invoice.DerivedLocale);
+
+            invoice.BillToCustomer = customer2;
+            this.Derive();
+
+            Assert.Equal(localeSe, invoice.DerivedLocale);
+        }
+
+        [Fact]
         public void ChangedAssignedCurrencyDeriveDerivedCurrency()
         {
             var invoice = new SalesInvoiceBuilder(this.Transaction).Build();
@@ -1670,6 +1698,25 @@ namespace Allors.Database.Domain.Tests
             var days = int.Parse(paymentDays.TermValue);
 
             Assert.Equal(invoice.DueDate, invoice.InvoiceDate.AddDays(days));
+        }
+
+        [Fact]
+        public void ChangedStorePaymentNetDaysDerivePaymentDays()
+        {
+            var store = new Stores(this.Transaction).Extent().First(v => Equals(v.InternalOrganisation, this.InternalOrganisation));
+            store.PaymentNetDays = 7;
+
+            var invoice = new SalesInvoiceBuilder(this.Transaction).WithStore(store).WithInvoiceDate(this.Transaction.Now()).Build();
+            this.Derive();
+
+            Assert.Equal(7, invoice.PaymentDays);
+            Assert.Equal(invoice.InvoiceDate.AddDays(7), invoice.DueDate);
+
+            store.PaymentNetDays = 14;
+            this.Derive();
+
+            Assert.Equal(14, invoice.PaymentDays);
+            Assert.Equal(invoice.InvoiceDate.AddDays(14), invoice.DueDate);
         }
 
         [Fact]